0.5.6 Minor Patch
Minor update patch to fix some errors and handle a few minor balance adjustments.
