+++ locale/en/castra.cfg
@@ -92,6 +92,7 @@ jammer-radar=Jammer radar
 jammed-data-collector=Jammed data collector
 castra-enemy-explosive-rc=Remote Explosive Car
 castra-enemy-car=Combat Buggy
+rc-car-explosion=Remote RC Detonation
 
 [entity-description]
 forge=Improves ammunition and weapon production.\nQuality: +10%
